requirements.txt
absl-py==1.1.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
anyio==3.5.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.2.0
asn1crypto==0.24.0
astroid==2.2.5
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.2.0
autobahn==20.12.3
Automat==20.2.0
autopep8==1.4.4
ax-platform==0.2.5.1
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.10.0
bidict==0.22.0
black==20.8b1
bleach==4.1.0
blinker==1.4
botorch==0.6.4
Brlapi==0.7.0
bson==0.5.10
cachetools==5.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.12
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
click==7.1.2
cloudinary==1.29.0
cmdtest==0.32+git
colorama==0.4.1
command-not-found==0.3
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.32
dbus-python==1.2.16
debugpy==1.6.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.13
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.2.24
django-allauth==0.40.0
django-countries==5.5
django-crispy-forms==1.7.2
django-tinymce4-lite==1.7.5
djangorestframework==3.12.4
dnspython==2.0.0
docopt==0.6.2
duplicity==0.8.12.0
ecdsa==0.18.0
email-validator==1.1.2
emnist==0.0
entrypoints==0.3
executing==0.8.3
fastapi==0.62.0
fastapi-sessions==0.3.2
fastapi-socketio==0.0.9
fasteners==0.14.1
filelock==3.8.0
flake8==3.8.4
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
fonttools==4.33.3
frozenlist==1.3.1
fsspec==2022.7.1
future==0.18.2
fvcore==0.1.5.post20220512
gast==0.4.0
GDAL==3.0.4
google-auth==2.8.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpytorch==1.8.1
greenlet==1.1.2
grpcio==1.43.0
h11==0.11.0
h2==3.2.0
h5py==3.7.0
hpack==3.0.0
html5lib==1.0.1
httpcore==0.15.0
httplib2==0.14.0
httptools==0.1.2
httpx==0.23.0
hydra-core==1.2.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
imageio==2.21.1
imgaug==0.4.0
importlib-metadata==4.12.0
importlib-resources==5.6.0
imutils==0.5.4
incremental==21.3.0
iopath==0.1.10
ipykernel==6.9.2
ipython==8.2.0
ipython-genutils==0.2.0
isort==4.3.21
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
jmespath==0.10.0
joblib==1.1.0
Js2Py==0.71
jsmin==2.2.2
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-server==1.15.6
jupyterlab==3.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.11.2
kaggle==1.5.12
keras==2.9.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.1
libclang==14.0.1
lmdb==1.3.0
lockfile==0.12.2
loguru==0.6.0
louis==3.12.0
lxml==4.7.1
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
motor==2.3.0
msgpack==1.0.4
multidict==6.0.2
multipledispatch==0.6.0
mypy-extensions==0.4.3
nbclassic==0.3.7
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
netifaces==0.10.4
networkx==2.8.6
ninja==1.10.2.3
nltk==3.7
notebook==6.4.10
notebook-shim==0.1.0
numpy==1.21.2
oauthlib==3.1.0
olefile==0.46
omegaconf==2.2.3
onnx==1.8.1
onnx-simplifier==0.3.5
onnxoptimizer==0.3.0
onnxruntime==1.8.0
opencv-python==4.5.5.64
opendatasets==0.1.22
opt-einsum==3.3.0
OWSLib==0.19.1
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
paramiko==2.6.0
parsel==1.6.0
parso==0.8.3
passlib==1.7.4
pathspec==0.8.1
pep8==1.7.1
pexpect==4.6.0
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==9.2.0
pipwin==0.5.2
plotly==4.4.1
portalocker==2.5.1
priority==1.3.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
Protego==0.1.16
protobuf==3.19.4
psutil==5.9.0
psycopg2==2.8.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycairo==1.16.2
pycocotools==2.0.4
pycodestyle==2.6.0
pycparser==2.19
pycups==1.9.73
pydantic==1.7.3
pyDeprecate==0.3.2
PyDispatcher==2.0.5
pyflakes==2.2.0
Pygments==2.11.2
PyGObject==3.36.0
PyHamcrest==1.9.0
pyjsparser==2.7.1
PyJWT==1.7.1
pylint==2.3.1
pymacaroons==0.13.0
pymongo==3.11.1
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPrind==2.11.3
pyproj==2.5.0
PyQt5==5.14.1
pyRFC3339==1.1
pyro-api==0.1.2
pyro-ppl==1.8.0
pyrsistent==0.18.1
pySmartDL==1.3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
python-engineio==4.3.4
python-jose==3.3.0
python-mnist==0.7
python-multipart==0.0.5
python-nmap==0.7.1
python-slugify==6.1.2
python-socketio==5.7.1
python3-openid==3.1.0
pytorch-lightning==1.6.5
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
pyxdg==0.26
PyYAML==6.0
pyzmq==22.3.0
queuelib==1.6.2
ray==1.13.0
razorpay==1.2.0
redis==4.3.4
regex==2020.11.13
reportlab==3.5.34
requests==2.28.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
retrying==1.3.3
rfc3986==1.5.0
roboflow==0.2.4
rsa==4.8
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.7.1
Scrapy==2.5.1
screen-resolution-extra==0.0.0
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.8.0
service-identity==21.1.0
Shapely==1.8.4
simplejson==3.16.0
sip==4.19.21
six==1.12.0
sklearn==0.0
sniffio==1.2.0
sounddevice==0.4.3
soupsieve==2.3.1
SQLAlchemy==1.4.39
sqlparse==0.2.4
stack-data==0.2.0
starlette==0.13.6
strhub==1.1.0
stripe==2.37.1
systemd-python==234
tabulate==0.8.10
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
text-unidecode==1.3
thop==0.1.1.post2207130030
threadpoolctl==3.1.0
tifffile==2022.8.12
timm==0.6.7
toml==0.10.2
torch==1.12.1
torchaudio==0.8.2
torchmetrics==0.9.3
torchvision==0.13.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
ttystatus==0.38
Twisted==21.7.0
txaio==21.2.1
typed-ast==1.4.1
typeguard==2.13.3
typing-extensions==3.7.4.3
tzdata==2022.1
tzlocal==4.2
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.6
usb-creator==0.3.7
uvicorn==0.12.3
uvloop==0.16.0
virtualenv==16.6.1
virtualenvwrapper-win==1.2.5
w3lib==1.22.0
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
websockets==10.3
Werkzeug==2.1.2
wget==3.2
wrapt==1.11.2
wsproto==0.15.0
xkit==0.0.0
yacs==0.1.8
yarl==1.8.1
yolox==0.3.0
zipp==3.7.0
zope.interface==5.4.0
